fixed PlaylistSong creation failing for pre-existing Songs
This was a regression, and just wasn't noticed because I just hadn't
happened to have tried to add a Song to multiple Playlists in a while.
The regression was introduced when we were refactoring the
implementation to allow creation of a Song together with a PlaylistSong.

If the Song exists, when we attempt to persist it via a PlaylistSong it
will throw up the constraint violation, but this was not the exception
we were seeing the in the stack. The exact exception was, "cannot save
transient entity", or such.

[TODOs]
- We will move back to REST semantics and have Song creation and
  PlaylistSong creation be two separate endpoints.