mv deduplication logic into SongService
While it was always a Song concern, we now have potentially two ways the
URL may be updated during lookups, so we need a single source of truth
for how to handle a Song update.

[fixes]
- #applyYouTubeResolution not actually updating the Song afterwards, so
  the resolution is never saved.
- potential unique constraint violations when applying the YouTube
  resolution. The URL may already be mapped in the system.
- missing guard clause for PlaylistSong#migration. This was picked up by
  DomainEntity#log's helpful thrown exception.

[revision]
- extracted SongMigration interface Provides a strategy for handling a Song
migration. This "decouples" the LookupService from the PlaylistSongService.
This is needed as we will soon introduce SongLookup, and eventually SongTag
which will both need to implement Migration strategies.